Move common strategy methods into own module
