+++ aquadoggo/src/replication/strategies/mod.rs
@@ -1,9 +1,11 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+mod common;
 mod diff;
 mod log_height;
 mod set_reconciliation;
 
+pub use common::{has_blob_relation, included_document_ids};
 pub use diff::diff_log_heights;
 pub use log_height::LogHeightStrategy;
 pub use set_reconciliation::SetReconciliationStrategy;
